default template with empty request-slots added. DE localization added
- added empty template (with settings to make it persistent and to configure its size)
- migration is updated
- player initialization is updated to take new settings into account
- DE localization added

+++ control.lua
@@ -54,8 +54,12 @@ script.on_event(defines.events.on_gui_click, function(event)
 		if preset_selected == 0 then
 			player.print({"messages.select-preset", {"messages.save"}})
 		else
-			request_manager.save_preset(player, preset_selected)
-			select_preset(player, preset_selected)
+			if request_manager.check_preset_protected (player, preset_selected, "save") == false then
+				request_manager.save_preset(player, preset_selected)
+				select_preset(player, preset_selected)
+			else
+				player.print ({"messages.protected-template", {"messages.overwritten"}})
+			end
 		end
 	
 	elseif gui_clicked == lrm.gui.load_button then
@@ -71,9 +75,13 @@ script.on_event(defines.events.on_gui_click, function(event)
 		if preset_selected == 0 then
 			player.print({"messages.select-preset", {"messages.delete"}})
 		else
-			request_manager.delete_preset(player, preset_selected)
-			gui.delete_preset(player, preset_selected)
-			select_preset(player, 0)
+			if request_manager.check_preset_protected (player, preset_selected) == false then
+				request_manager.delete_preset(player, preset_selected)
+				gui.delete_preset(player, preset_selected)
+				select_preset(player, 0)
+			else
+				player.print ({"messages.protected-template", {"messages.deleted"}})
+			end
 		end
 	
 	else
@@ -100,21 +108,21 @@ script.on_event(defines.events.on_player_created, function(event)
 	local player = game.players[event.player_index]
 	if not (player and player.valid) then return end
 	
-	globals.init_player(player)
-	gui.build(player)
+	init_player (player)
 end)
 
 script.on_init(function()
 	globals.init()
-
 	for _, player in pairs(game.players) do
-		globals.init_player(player)
-		gui.build(player)
+		init_player (player)
 	end
 end)
 
 script.on_configuration_changed(function()
 	globals.init()
+	for _, player in pairs(game.players) do
+		init_player (player)
+	end
 end)
 
 
@@ -122,9 +130,37 @@ script.on_event(defines.events.on_runtime_mod_setting_changed,function(event)
 	player = game.players[event.player_index]
 	setting = event.setting
 	value = event.value
-	game.print("on_runtime_mod_setting_changed: " .. serpent.dump(event))
+	get_player_setting (player, setting)
 end)
 
-function read_settings ()
+function get_player_setting (player, setting)
+	if not (player) then return nil end
+	if not (setting) then return nil end
+	
+	local value = game.players[player.index].mod_settings[setting].value
+	local old_value = global["player_settings"][player.index][setting]
+	global["player_settings"][player.index][setting] = value
+	if not ( value == old_value) then
+		if ( setting==lrm.settings.empty_template_size ) or ( ( setting==lrm.settings.persistent_empty_template ) and value==true ) then
+			request_manager.create_empty_template(player)
+			gui.force_rebuild(player)
+			local preset=global["presets-selected"][player.index] or 0
+			if preset == 0 then preset = 1 end
+			select_preset(player, preset)
+		end
+	end
+	return (value)
+end
 
+function init_settings (player)
+	if not (player) then return nil end
+	get_player_setting (player, lrm.settings.persistent_empty_template)
+	get_player_setting (player, lrm.settings.empty_template_size)
+end
+
+function init_player (player)
+	globals.init_player(player)
+	init_settings (player)
+	request_manager.create_empty_template(player)
+	gui.build(player)
 end
